+++ CHANGES.rst
@@ -12,6 +12,19 @@ Changes
 
 Changes:
 --------
+- No change.
+
+Fixes:
+------
+- No change.
+
+.. _changes_5.9.0:
+
+`5.9.0 <https://github.com/crim-ca/weaver/tree/5.9.0>`_ (2024-09-12)
+========================================================================
+
+Changes:
+--------
 - Add `CWL` schema definitions with ``weaver`` namespace
   (see `weaver/schemas/cwl <https://github.com/crim-ca/weaver/tree/master/weaver/schemas/cwl>`_)
   that provide explicit requirement classes
@@ -30,6 +43,9 @@ Changes:
   and a sample `crim-ca/ncml2stac <https://github.com/crim-ca/ncml2stac/tree/main#ncml-to-stac>`_ repository
   making use of it with the `Weaver` `CLI` to generate a deployed `OGC API - Processes` definition
   (fixes `#63 <https://github.com/crim-ca/weaver/issues/63>`_).
+- Add parsing of additional metadata from schema.org in CWL document to convert into process fields
+  (fixes `#463 <https://github.com/crim-ca/weaver/issues/463>`_).
+- Add more metadata mapping details in documentation (fixes `#613 <https://github.com/crim-ca/weaver/issues/613>`_).
 
 Fixes:
 ------
@@ -38,6 +54,9 @@ Fixes:
   disallow ``additionalProperties`` that cannot be mapped to a particular child `JSON` schema definition.
 - Fix ``VariableSchemaNode`` resolution to allow mapping against multiple ``variable`` sub-nodes representing
   different nested `JSON` schema nodes permitted under the ``additionalProperties`` mapping.
+- Fix ``GET /jobs`` endpoint failing to return the rendered `HTML` listing when ``detail=true`` was omitted or
+  set to any non-detailed value. The ``detail`` query parameter is ignored for `HTML` since details are always
+  required to populate the `Job` table.
 - Pin ``pymongo>=4.3`` and remove ``celery[mongodb]`` extra requirement to avoid incompatible resolution
   of ``pymongo[srv]>=4.8.0`` (relates to `celery/celery#9254 <https://github.com/celery/celery/issues/9254>`_
   and `MongoDB PYTHON-4756 <https://jira.mongodb.org/browse/PYTHON-4756>`_).